+++ app/components/common/waline.tsx
@@ -1,5 +1,6 @@
 import { init } from "@waline/client";
 import "@waline/client/style";
+import { walinePath } from "lib/waline-path";
 import { useEffect } from "react";
 import { siteInfo } from "site.config";
 import "./waline.scss";
@@ -14,7 +15,7 @@ const Waline = (props: React.HTMLProps<HTMLDivElement>) => {
     init({
       el: "#waline",
       serverURL: siteInfo.walineApi,
-      path: window.location.pathname,
+      path: walinePath(window.location.pathname),
       pageview: true,
       comment: true,
     });